+++ src/pages/signUp/SignUp.tsx
@@ -10,7 +10,7 @@ import { ButtonComponent } from '../../components/button/ButtonComponent';
 import { SnackBar } from '../../components/snackBar/SnackBar';
 import { Wrapper } from '../../components/wrapper/Wrapper';
 import { signUp } from '../../store/thunks/thunks';
-import { useAppDispatch, useAppSelector } from '../../utils/hooks/hooks';
+import { useAppDispatch, useAppSelector } from '../../utils/hooks/useSelectorUseDispatch';
 import { RegistrationType, SignUpType } from '../../utils/types/types';
 import { signUpValidate } from '../../utils/validators/validators';
