+++ src/utils/constants.ts
@@ -4,6 +4,7 @@ export enum path {
   forgot = '/forgot',
   checkEmail = '/check-email',
   newPassword = '/set-new-password',
+  packsList = '/packsList',
 }
 
 export const minPasswordLength = 8;
